Basic Poll Model and Tests
Creates the basic poll model and associated tests. The poll model
contains a required question and owner, everything else is optional.
The owner is an objectid which points to the _id of an owner in the
User collection. The model contains a couple helper functions to
return the associated user object as well as vote for specific
poll options and adding an option to the polls (with initially 0
votes).