Install Bootstrap and JQuery
Installs bootstrap and jQuery and links them to express static
middleware /bootstrap and /jquery. Adds a new route for the
signup page to test whether a username exists or not and adds two
tests for this new route. Updates the signup page to utilize
bootstrap and jQuery. Additionally, adds an express static
middleware linked to /scripts to hold javascript files. New javascript
file created in /scripts to utilize jQuery and add form validation
as well as the username check. Current placeholders set up.