Ability to Vote for Different Polls
Creates the ability for a user to vote for different polls. This
commit includes a new POST path to /posts/:id/vote and will read the
body of the post request to determine which option was voted for. This
is then used in the poll model VoteFor function. This function was
modified to no longer return the new poll object in the callback as
this is already available. The addOption function was also changed to
reflect this. Includes modified tests for the poll model and new tests
for the POST route.