Adds ChartJS Display of Poll and Voting Issue
Installs and adds a ChartJS display for each individual poll. The
poll will be displayed on /polls/:id. Fixes an issue with voting
for new options as a logged in user. When an authenticated user
added an option, uncaught errors caused the application to crash.
Adds better error checking to prevent crashes. Adds two tests for
voting for a poll and voting for a new option as an authenticated
user.