View Polls With Paging Feature
Creates the /polls/ route and allows users to view a list of polls
sorted by newest poll first. The skip parameter in the URL allows
users to navigate the list by viewing 25 polls at a time. Adds the
User.populate to populate the username for displaying the polls
before returning to the template.