Adds Twitter Login Functionality
Installs appropriate packages for logging in via Twitter. Creates
a new Twitter Authentication Strategy and adds to the config file
the environment variables for the API key and Secret. Adds a new
User model static function to assist with logging in a Twitter
user. If one exists, it will log them in, otherwise, it will create
a new user and then log them in. Tests created for the new model
method.